refactor: add module(:core:model)
