+++ app/src/main/java/kosenda/makecolor/model/ColorDao.kt
@@ -4,6 +4,7 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import kosenda.makecolor.model.data.ColorItem
 
 @Dao
 interface ColorDao {
