+++ app/src/main/java/kosenda/makecolor/model/ColorDatabase.kt
@@ -6,6 +6,7 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import androidx.room.migration.Migration
 import androidx.sqlite.db.SupportSQLiteDatabase
+import kosenda.makecolor.model.data.ColorItem
 
 @Database(
     version = 2,
