+++ app/src/main/java/kosenda/makecolor/view/screen/SelectColorScreen.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.google.accompanist.systemuicontroller.rememberSystemUiController
 import kosenda.makecolor.R
-import kosenda.makecolor.model.ColorItem
+import kosenda.makecolor.model.data.ColorItem
 import kosenda.makecolor.model.util.getNameIfNoAlias
 import kosenda.makecolor.view.PreviewSurface
 import kosenda.makecolor.view.component.button.TonalButton
