+++ app/src/main/java/kosenda/makecolor/view/state/CategoryDetailUiState.kt
@@ -1,7 +1,7 @@
 package kosenda.makecolor.view.state
 
 import kosenda.makecolor.model.Category
-import kosenda.makecolor.model.ColorItem
+import kosenda.makecolor.model.data.ColorItem
 import kosenda.makecolor.model.default.webColor
 
 data class CategoryDetailUiState(
