+++ app/src/main/java/kosenda/makecolor/view/state/MergeUiState.kt
@@ -2,7 +2,7 @@ package kosenda.makecolor.view.state
 
 import android.graphics.Bitmap
 import kosenda.makecolor.model.Category
-import kosenda.makecolor.model.ColorData
+import kosenda.makecolor.model.data.ColorData
 import kosenda.makecolor.model.util.randomColorData
 import kosenda.makecolor.model.util.randomHex
