+++ app/src/main/java/kosenda/makecolor/view/state/PictureUiState.kt
@@ -1,7 +1,7 @@
 package kosenda.makecolor.view.state
 
 import android.graphics.Bitmap
-import kosenda.makecolor.model.ColorData
+import kosenda.makecolor.model.data.ColorData
 import kosenda.makecolor.model.util.randomColorData
 
 data class PictureUiState(
