+++ app/src/main/java/kosenda/makecolor/view/state/RandomUiState.kt
@@ -1,7 +1,7 @@
 package kosenda.makecolor.view.state
 
-import kosenda.makecolor.model.ColorData
-import kosenda.makecolor.model.RGB
+import kosenda.makecolor.model.data.ColorData
+import kosenda.makecolor.model.data.RGB
 import kosenda.makecolor.model.util.outputRandomRGBColors
 import kosenda.makecolor.model.util.randomColorData
 import kosenda.makecolor.view.code.RandomType
