+++ app/src/main/java/kosenda/makecolor/viewmodel/PreviewViewModel.kt
@@ -11,7 +11,6 @@ import androidx.navigation.NavBackStackEntry
 import com.godaddy.android.colorpicker.HsvColor
 import com.godaddy.android.colorpicker.harmony.ColorHarmonyMode
 import kosenda.makecolor.model.Category
-import kosenda.makecolor.model.ColorItem
 import kosenda.makecolor.model.StringResource
 import kosenda.makecolor.model.default.webColor
 import kosenda.makecolor.view.CMYKColor
@@ -139,7 +138,7 @@ class PreviewSelectColorViewModel : SelectColorViewModel() {
         SelectColorUiState(colors = webColor),
     )
     override fun getColors() {}
-    override fun deleteColorItem(deleteColorItem: ColorItem) {}
+    override fun deleteColorItem(deleteColorItem: kosenda.makecolor.model.data.ColorItem) {}
 }
 
 class PreviewSettingsViewModel : SettingsViewModel() {
