+++ core/model/src/main/java/kosenda/makecolor/model/data/ColorItem.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor.model
+package kosenda.makecolor.model.data
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
