+++ gradle/libs.versions.toml
@@ -1,10 +1,10 @@
 [versions]
 accompanist = "0.28.0"
-androidGradlePlugin = "7.4.2"
+androidGradlePlugin = "7.4.1"
 androidxActivity = "1.6.1"
 androidxAppCompat = "1.7.0-alpha02"
-androidxComposeCompiler = "1.3.1"
-androidxComposeMaterial = "1.3.1"
+androidxComposeCompiler = "1.4.0"
+androidxComposeMaterial = "1.4.0"
 androidxComposeMaterial3 = "1.1.0-beta01"
 androidxCore = "1.9.0"
 androidxCoreSplashscreen = "1.0.0"
@@ -21,10 +21,10 @@ firebaseBom = "31.1.1"
 gmsPlugin = "4.3.15"
 hilt = "2.45"
 junit4 = "4.13.2"
-kotlin = "1.7.10"
+kotlin = "1.8.0"
 kotlinJdk = "1.8.0"
 kotlinxSerializationJson = "1.4.1"
-ksp = "1.7.10-1.0.6"
+ksp = "1.8.0-1.0.9"
 ktlint = "0.48.0"
 navigationAnimation = "0.24.5-alpha"
 ossLicenses = "17.0.0"
